Nagardocs-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
pydantic-settings
python-dotenv

# Auth
passlib[bcrypt]
python-jose[cryptography]

# HTTP (Supabase REST)
httpx

# File upload
python-multipart
aiofiles

# OCR
pytesseract
Pillow
pymupdf

# AI
openai

# Security / rate limit
slowapi

# Data export
pandas
openpyxl
python-dateutil
